+++ NEWS.md
@@ -1,4 +1,5 @@
 # eatATA 1.0.0.9000
+* bugfix in `inspectSolution()` (works with `tibbles` now)
 * bugfix in `itemTuples()` (#1)
 
 # eatATA 1.0.0